rbd: flatten the clone image if pools are different
currently we flatten all the clones once we reach
the maxSnapshotsOnImage count, we assumes that the clones
are created in same pool with same name as snapshot,as
we are supporting the creation of clone in different pool,
its not possible to get the pool name in which all the
clones are created, if the request is to create a clone
in a different pool. create the clone and flatten the
clone so that clone chain will be broken.

Signed-off-by: Madhu Rajanna <madhupr007@gmail.com>