doc: cleanup Pull Request templates
GitHub does not handle multiple Pull Request templates the same as it
does for Issue templates. When creating a PR it is required to pass
`?template=<filename>` on the URL to select a template.

To make the URLs a little shorter and simpler, the templates have been
renamed to drop the `00-` prefix. Ordering of the templates is not
needed, as they do not show up in a Pull Request menu anyway.

The additional header that Issue templates have are not used with Pull
Request templates either, so these have been removed as well.

Signed-off-by: Niels de Vos <ndevos@redhat.com>