e2e: introducing an interface for deployment of rbd and cephfs
This commit introduce a new common interface for deployment of
rbd and cephfs deployments in e2e. As part of this a new file
called e2e/deployer.go has been introduced with below methods
in it.

type deployer interface {
	replaceNamespaceInTemplate(string) (string, error)
	runAction(string, string, string) error
}

These methods handle common tasks for the deployment process of
RBD and CephFS.The individual deployers are free to override the methods
or extend their own later.
For example:
                changeTemplateAndRunAction(string, string)

is 'deployer' specific ( rbd, cephfs...etc) which handle deployer specific
template changes and cases.

Signed-off-by: Humble Chirammal <hchiramm@redhat.com>