e2e: make it possible to disable metrics verification
By introducing --test-metrics (defaults to 'true') it is now possible to
disable testing of the metrics per PVC. This is enabled by default when
Kubernetes 1.22 or newer is used, older versions do not expose metrics
of BlockMode volumes through Kubelet.

In some deployments (like OpenShift) it is not possible to access the
Kubelet metrics directly. Disabling the verification of the metrics
makes it possible to run the e2e tests on those deployments too.

Signed-off-by: Niels de Vos <ndevos@redhat.com>