e2e: allow deploying Vault Pod on OpenShift
The UID of the user can not be used in OpenShift, it needs to be much
higher (> 1000000?). Dropping the `runAs` automatically picks a random
UID.

On OpenShift the root (`/`) directory is not writable in the container.
This causes a failure while staring the service. Setting the
home-directory of the user to `/tmp` makes deploying work.

Signed-off-by: Niels de Vos <ndevos@redhat.com>