e2e: count number of RBD images that are available in the pool before testing
When the e2e tests are run on a deployment that already has RBD images,
the tests will fail because the expected number of images is not
correct. By counting the number of images before the tests start, the
expected number of images can be verified.

This still expects that the pool is modified outside of the e2e tests
while testing is in progress.

Signed-off-by: Niels de Vos <ndevos@redhat.com>